requirements.txt
torch==1.13.1
torchvision==0.14.1
tensorflow==2.11.0
ibm-cos-sdk==2.12.2
torch-model-archiver>=0.7.1
Flask==2.2.3
celery[redis]==5.2.7
apispec==6.2.0
apispec-webframeworks==0.5.2
marshmallow==3.19.0
onnx2pytorch
onnx
tf2onnx
openshift
kserve
kubernetes
tqdm
python-dotenv
torchextractor
ultralytics
transformers==4.24.0
datasets==2.13.1
numpy==1.23.5
